requirements.txt
flask
flask-sqlalchemy
wheel
psycopg2-binary
kafka-python